refactor(CANCEL-1): collapse the two cancellation lanes into one implementation
The permission and question cancellation directories had drifted into ~580 lines
of near-identical logic differing only by configuration: request lane, response
lane, in-flight kind, parser, handler, missing-handler label and log label. Every
fix in this area had to be written twice, which is how the lanes disagreed about
retry semantics in the first place.

One lane-parameterised implementation now takes those seven inputs; both exported
entry points remain as thin wrappers with unchanged signatures, so ipc.ts is
untouched. The lane types are Extract-constrained, so a wrong lane fails to
typecheck rather than failing at runtime. Durable-record state stays lane-scoped
via the existing lane-qualified records directory.

Net -520 lines, behaviour-neutral: the describe.each suite covering both lanes
passes completely unmodified, which was the precondition for doing this at all.

Refs: docs/decisions/0056-durable-cancellation-invariant.md
