+++ apps/core/src/runtime/ipc-question-cancellation-directory.ts
@@ -1,34 +1,12 @@
-import fs from 'fs';
-import path from 'path';
-import { createHash } from 'crypto';
-
-import type { UserQuestionCancellation } from '../domain/types.js';
-import { IPC_CANCELLATION_RETENTION_TTL_MS } from '../shared/ipc-cancellation-lifetime.js';
+import { parseQuestionCancellationIpcRequest } from './ipc-parsing.js';
 import type { IpcDeps } from './ipc-domain-types.js';
 import type { FilesystemRunnerControlPort } from './filesystem-runner-control-port.js';
 import {
-  claimDurableCancellationRecord,
-  createDurableCancellationRecord,
-  durableCancellationRecordsDir,
-  listDurableCancellationRecords,
-  readDurableCancellationRecord,
-  releaseDurableCancellationRecord,
-  type DurableCancellationRecord,
-} from './ipc-cancellation-durable-record.js';
-import { interactionInFlightKey } from './ipc-interaction-processing.js';
-import { parseQuestionCancellationIpcRequest } from './ipc-parsing.js';
+  processCancellationDirectory,
+  type CancellationDirectoryLogger,
+} from './ipc-cancellation-directory.js';
 
 const QUESTION_CANCELLATION_LANE = 'question-cancellations';
-const CANCELLATION_RETRY_MIN_MS = 1_000;
-const CANCELLATION_RETRY_MAX_MS = 30_000;
-
-type CancellationRetryState =
-  DurableCancellationRecord<UserQuestionCancellation>;
-
-type QuestionCancellationDirectoryLogger = {
-  error(context: Record<string, unknown>, message: string): void;
-  warn(context: Record<string, unknown>, message: string): void;
-};
 
 export async function processQuestionCancellationDirectory(input: {
   sourceAgentFolder: string;
@@ -39,251 +17,15 @@ export async function processQuestionCancellationDirectory(input: {
   inFlightInteractionIpc: ReadonlySet<string>;
   runnerControlPort: FilesystemRunnerControlPort;
   cancelUserQuestion: IpcDeps['cancelUserQuestion'];
-  logger: QuestionCancellationDirectoryLogger;
+  logger: CancellationDirectoryLogger;
 }): Promise<void> {
-  const { sourceAgentFolder, runnerControlPort, logger } = input;
-  const cancellationsDir = runnerControlPort.requestDir(
-    sourceAgentFolder,
-    QUESTION_CANCELLATION_LANE,
-  );
-  try {
-    if (
-      !input.shouldProcessRequestLane(
-        sourceAgentFolder,
-        QUESTION_CANCELLATION_LANE,
-      ) ||
-      !runnerControlPort.isTrustedRequestDir(
-        sourceAgentFolder,
-        QUESTION_CANCELLATION_LANE,
-      )
-    ) {
-      return;
-    }
-    const files = runnerControlPort.listPendingRequests(
-      sourceAgentFolder,
-      QUESTION_CANCELLATION_LANE,
-    );
-    const recordsDir = durableCancellationRecordsDir(
-      runnerControlPort.baseDir,
-      QUESTION_CANCELLATION_LANE,
-      sourceAgentFolder,
-    );
-    const retryFiles = listDurableCancellationRecords(recordsDir);
-    const now = Date.now();
-    for (const file of files) {
-      await ingestQuestionCancellationFile(
-        input,
-        cancellationsDir,
-        recordsDir,
-        file,
-      );
-    }
-    for (const file of retryFiles) {
-      const retry = loadPendingCancellationRetryState(recordsDir, file);
-      if (retry && retry.nextAttemptAt > now && retry.expiresAt > now) {
-        continue;
-      }
-      await processQuestionCancellationRecord(input, recordsDir, file);
-    }
-  } catch (err) {
-    logger.error(
-      { err, sourceAgentFolder },
-      'Error reading question cancellation IPC requests directory',
-    );
-  }
-}
-
-async function ingestQuestionCancellationFile(
-  input: Parameters<typeof processQuestionCancellationDirectory>[0],
-  cancellationsDir: string,
-  recordsDir: string,
-  file: string,
-): Promise<void> {
-  const { sourceAgentFolder, runnerControlPort, logger } = input;
-  let claimedPath = path.join(cancellationsDir, file);
-  let recordFile: string | undefined;
-  try {
-    const claimed = runnerControlPort.claimRequest(
-      sourceAgentFolder,
-      QUESTION_CANCELLATION_LANE,
-      file,
-    );
-    claimedPath = claimed.claimedPath;
-    const envelopeDigest = cancellationEnvelopeDigest(claimed.raw);
-    const cancellation = parseQuestionCancellationIpcRequest(
-      claimed.raw,
-      sourceAgentFolder,
-    );
-    const now = Date.now();
-    recordFile = createDurableCancellationRecord(recordsDir, {
-      attempts: 0,
-      cancellation,
-      envelopeDigest,
-      expiresAt:
-        Math.min(now, fs.statSync(claimedPath).mtimeMs) +
-        IPC_CANCELLATION_RETENTION_TTL_MS,
-      nextAttemptAt: now,
-    });
-    fs.unlinkSync(claimedPath);
-  } catch (err) {
-    logger.error(
-      { file, sourceAgentFolder, err },
-      'Error processing question cancellation IPC request',
-    );
-    runnerControlPort.archiveFailedRequest(
-      sourceAgentFolder,
-      file,
-      claimedPath,
-    );
-    return;
-  }
-  await processQuestionCancellationRecord(input, recordsDir, recordFile);
-}
-
-async function processQuestionCancellationRecord(
-  input: Parameters<typeof processQuestionCancellationDirectory>[0],
-  recordsDir: string,
-  file: string,
-): Promise<void> {
-  const { sourceAgentFolder, runnerControlPort, logger } = input;
-  let claimedPath = path.join(recordsDir, file);
-  try {
-    claimedPath = claimDurableCancellationRecord(recordsDir, file);
-    const retry =
-      readDurableCancellationRecord<UserQuestionCancellation>(claimedPath);
-    const cancellation = retry.cancellation;
-    if (!isQuestionInFlight(input.inFlightInteractionIpc, cancellation)) {
-      if (
-        runnerControlPort.responseExists(
-          sourceAgentFolder,
-          'user-answers',
-          cancellation.requestId,
-        )
-      ) {
-        fs.unlinkSync(claimedPath);
-        return;
-      }
-      retainCancellation({
-        claimedPath,
-        recordsDir,
-        file,
-        logger,
-        sourceAgentFolder,
-        retry,
-      });
-      return;
-    }
-    if (!input.cancelUserQuestion) {
-      throw new Error('Question cancellation handler is unavailable');
-    }
-    let result: Awaited<ReturnType<NonNullable<IpcDeps['cancelUserQuestion']>>>;
-    try {
-      result = await input.cancelUserQuestion(cancellation);
-    } catch (err) {
-      logger.error(
-        { file, sourceAgentFolder, err },
-        'Error processing question cancellation IPC request',
-      );
-      retainCancellation({
-        claimedPath,
-        recordsDir,
-        file,
-        logger,
-        sourceAgentFolder,
-        retry,
-      });
-      return;
-    }
-    if (result === 'settled') {
-      fs.unlinkSync(claimedPath);
-      return;
-    }
-    retainCancellation({
-      claimedPath,
-      recordsDir,
-      file,
-      logger,
-      sourceAgentFolder,
-      retry,
-    });
-  } catch (err) {
-    logger.error(
-      { file, sourceAgentFolder, err },
-      'Error processing question cancellation IPC request',
-    );
-    runnerControlPort.archiveFailedRequest(
-      sourceAgentFolder,
-      file,
-      claimedPath,
-    );
-  }
-}
-
-function isQuestionInFlight(
-  inFlightInteractionIpc: ReadonlySet<string>,
-  cancellation: UserQuestionCancellation,
-): boolean {
-  return inFlightInteractionIpc.has(
-    interactionInFlightKey({
-      sourceAgentFolder: cancellation.sourceAgentFolder,
-      kind: 'user-question',
-      threadId: cancellation.threadId,
-      requestId: cancellation.requestId,
-    }),
-  );
-}
-
-function retainCancellation(input: {
-  claimedPath: string;
-  recordsDir: string;
-  file: string;
-  logger: QuestionCancellationDirectoryLogger;
-  sourceAgentFolder: string;
-  retry: CancellationRetryState;
-}): void {
-  const now = Date.now();
-  if (input.retry.expiresAt <= now) {
-    fs.unlinkSync(input.claimedPath);
-    input.logger.warn(
-      {
-        file: input.file,
-        sourceAgentFolder: input.sourceAgentFolder,
-        retentionMs: IPC_CANCELLATION_RETENTION_TTL_MS,
-      },
-      'Discarding expired question cancellation IPC request',
-    );
-    return;
-  }
-
-  const attempts = input.retry.attempts + 1;
-  const retryDelayMs = Math.min(
-    CANCELLATION_RETRY_MIN_MS * 2 ** (attempts - 1),
-    CANCELLATION_RETRY_MAX_MS,
-  );
-  releaseDurableCancellationRecord(
-    input.claimedPath,
-    path.join(input.recordsDir, input.file),
-    {
-      ...input.retry,
-      attempts,
-      nextAttemptAt: now + retryDelayMs,
-    },
-  );
-}
-
-function cancellationEnvelopeDigest(raw: unknown): string {
-  return createHash('sha256').update(JSON.stringify(raw)).digest('hex');
-}
-
-function loadPendingCancellationRetryState(
-  recordsDir: string,
-  file: string,
-): CancellationRetryState | undefined {
-  try {
-    return readDurableCancellationRecord<UserQuestionCancellation>(
-      path.join(recordsDir, file),
-    );
-  } catch {
-    return undefined;
-  }
+  return processCancellationDirectory(input, {
+    requestLane: QUESTION_CANCELLATION_LANE,
+    responseLane: 'user-answers',
+    inFlightKind: 'user-question',
+    parser: parseQuestionCancellationIpcRequest,
+    handler: input.cancelUserQuestion,
+    missingHandlerErrorLabel: 'Question cancellation',
+    logLabel: 'question cancellation',
+  });
 }
